+++ src/components/Modals/AproveReceiptsModal.tsx
@@ -3,10 +3,10 @@
  * in the accounts payable module.
  */
 
-import { useCallback } from "react";
+import { useCallback, useState } from "react";
 import { apiRequest } from "@utils/apiClient";
 import ModalWrapper from "@/components/Modals/ModalWrapper";
-import Button from "@/components/Buttons/Button";
+import Toast from '@/components/Utils/Toast';
 
 interface Props {
   receipt_id: number;
@@ -14,8 +14,9 @@ interface Props {
   message: string;
   redirection: string;
   modal_type: "success" | "warning";
+  color?: "success" | "warning" | "primary" | "secondary";
   variant?: "filled" | "border" | "empty";
-  children: React.ReactNode;
+  label?: string;
   disabled?: boolean;
   token: string;
 }
@@ -26,50 +27,46 @@ export default function ApproveReceiptStatus({
   message,
   redirection,
   modal_type,
-  variant,
-  children,
+  color = "success",
+  variant = "filled",
+  label = "Aprobar",
   disabled = false,
   token
 }: Props) {
+  const [toast, setToast] = useState<{ message: string; type: 'success' | 'error' } | null>(null);
   const handleConfirm = useCallback(async () => {
     try {
-        const url = `/accounts-payable/validate-receipt/${receipt_id}`;
+      const url = `/accounts-payable/validate-receipt/${receipt_id}`;
       await apiRequest(url, { 
         method: "PUT", 
         data: {"approval": 1},
         headers: { Authorization: `Bearer ${token}` }
       });
-      alert(`Aprobado correctamente`)
+      setToast({ message: 'Comprobante aprobado exitosamente.', type: 'success' });
 
-      if (redirection) {
+      setTimeout(() => {
         window.location.reload();
-      } else {
-        window.location.reload();
-      }
+      }, 2000);
     } catch (error) {
       console.error("Error en la solicitud:", error);
+      setToast({ message: 'Error al aprobar el comprobante.', type: 'error' });
     }
   }, [receipt_id, redirection]);
 
   return (
-    <ModalWrapper
-      title={title}
-      message={message}
-      color="success"
-      modal_type={modal_type}
-      onConfirm={handleConfirm}
-      disabled={disabled}
-      triggerElement={
-        <Button
-          color="success"
-          variant="filled"
-          size="medium"
-          disabled={disabled}
-          customSizeClass="w-full"
-        >
-          {children}
-        </Button>
-      }
-    />
+    <>
+      {toast && <Toast message={toast.message} type={toast.type} />}
+      <ModalWrapper
+        title={title}
+        message={message}
+        color={color}
+        modal_type={modal_type}
+        variant={variant}
+        disabled={disabled}
+        onConfirm={handleConfirm}
+      >
+        {label}
+      </ModalWrapper>
+    </>
   );
 }